refactor: 모달 렌더링 조건 변경
