+++ src/components/restaurant/RestaurantList.tsx
@@ -1,12 +1,10 @@
 import { Component } from 'react';
 import styled from 'styled-components';
 import SelectBox from '../common/SelectBox';
-import mockData from '../../mockData.json';
 import RestaurantItem from './RestaurantItem';
 import restaurant from '../../domain/restaurant';
-import { getLocalStorage, setLocalStorage } from '../../utils/localStorage';
-import { CATEGORIES, SORT_OPTIONS, LOCAL_STORAGE_KEY } from '../../constants';
-import { Restaurant, SetModalRestaurant } from '../../@types/type';
+import { CATEGORIES, SORT_OPTIONS } from '../../constants';
+import { Restaurant, SetModalRestaurantId } from '../../@types/type';
 
 const RestaurantListLayout = styled.main`
   padding: 16px;
@@ -23,26 +21,19 @@ const Restaurants = styled.ul`
   margin: 16px 0;
 `;
 
-class RestaurantList extends Component<SetModalRestaurant> {
+type Props = SetModalRestaurantId & {
+  restaurantList: Restaurant[];
+};
+
+class RestaurantList extends Component<Props> {
   state = {
-    restaurantList: mockData as Restaurant[],
     filterOption: CATEGORIES.ALL,
     sortOption: SORT_OPTIONS.NAME,
   };
 
-  componentDidMount() {
-    const savedRestaurants = getLocalStorage(LOCAL_STORAGE_KEY.RESTAURANT);
-
-    if (savedRestaurants) {
-      this.setState({
-        restaurantList: savedRestaurants,
-      });
-    } else setLocalStorage(LOCAL_STORAGE_KEY.RESTAURANT, this.state.restaurantList);
-  }
-
   filterAndSort = () => {
     return restaurant.sort(
-      restaurant.filter(this.state.restaurantList, this.state.filterOption),
+      restaurant.filter(this.props.restaurantList, this.state.filterOption),
       this.state.sortOption,
     );
   };
@@ -68,7 +59,11 @@ class RestaurantList extends Component<SetModalRestaurant> {
         </SelectBoxContainer>
         <Restaurants>
           {this.filterAndSort().map((restaurant, index) => (
-            <RestaurantItem key={index} restaurant={restaurant} setModalRestaurant={this.props.setModalRestaurant} />
+            <RestaurantItem
+              key={index}
+              restaurant={restaurant}
+              setModalRestaurantId={this.props.setModalRestaurantId}
+            />
           ))}
         </Restaurants>
       </RestaurantListLayout>
